Move unfold2d. Update vision_benchmark
